Revert to use one centralized config file under /internal/config/config.go
