+++ backend/internal/auth/service.go
@@ -11,6 +11,8 @@ import (
 
 	"github.com/coreos/go-oidc/v3/oidc"
 	"golang.org/x/oauth2"
+
+	"hkers-backend/internal/config"
 )
 
 // Service handles authentication logic with a generic OIDC provider.
@@ -21,11 +23,11 @@ type Service struct {
 	clientID      string
 	endSessionURL string
 	postLogoutURL string
-	oidcCfg       OIDCConfig
+	oidcCfg       config.OIDCConfig
 }
 
 // NewService creates a new OIDC authentication service instance.
-func NewService(cfg *OIDCConfig) (*Service, error) {
+func NewService(cfg *config.OIDCConfig) (*Service, error) {
 	// Validate required configuration
 	required := []struct {
 		value string
